chore: Update environment configuration files and npm dependencies
This commit updates the environment configuration files for both the client and server. It also updates the npm dependencies and Firebase configuration. Additionally, the .gitignore file has been updated to ignore the .env file and the Server/node_modules directory has been deleted. These changes improve the security and organization of the repository.